Fully functional stand-alone version for static scenarios, using DerbyDB
-- Added class DerbyResults that re-implements all the functionality in
class MongoResults, using the Derby database. 
-- Made required changes in Simulation, in order for class DerbyResults
to be used, instead of MongoResults.

-- Commented out non-stand-alone versions of methods for setting up and
running scenarios in Simulation.

-- Minor corrections in StandAloneSimulation to properly formulate the
"comparison" scenario (scenario to be run both on the CASSANDRA platform
and using the stand-alone version of the code).

-- Cleaned up code. Deleted unused classes Runs, ApplianceTest,
ConsumptionModelTest, GaussianCDFTest, GaussianMixtureModelsTest,
GaussianTest, UniformTest, EventTest, RestfulTestProject,
SimCalendarTest & SimulationTest.

-- Added classes WwdEmbedded & WWdUtils that contain a basic DerbyDB
example. (These classes are to be removed later).

-- Minor clean-up of code in classes Installation & Appliance.